Improves organization of different architectures
Now each architecture has its own batch function. Further, the non dag architectures now have a batch function which automatically converts the elements of the masks to {-1,1}, and also decides if the otuput batch should be a gpuArray or not. cnn_train was also changed so that work is not done there anymore.